Better error when required package is not installed
Using an operator that requires a package that is not installed results in
an error.  Before this commit, the error message did not clearly describe
the problem.  For example, trying to use evil-highlight-operator without
the highlight package installed resulted in the following error:

    Cannot open load file: No such file or directory, highlight

This commit improves error messages to more clearly indicate the problem
and its solution:

    You must install the highlight package to use evil-operator-highlight

* evil-extra-operator.el (evil-operator-google-translate):
(evil-operator-highlight):
(evil-operator-fold):
(evil-operator-org-capture):
(evil-operator-remember): If a required package is not found, tell the
user that the package needs to be installed.